Rearrange SqlStatement files
