+++ source/Databricks/source/SqlStatementExecution.Tests/Models/SqlResponseTests.cs
@@ -12,11 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Internal.Models;
+using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Models;
 using Energinet.DataHub.Core.TestCommon.AutoFixture.Attributes;
 using Xunit.Categories;
 
-namespace Energinet.DataHub.Core.Databricks.SqlStatementExecution.Tests.Internal.Models;
+namespace Energinet.DataHub.Core.Databricks.SqlStatementExecution.Tests.Models;
 
 [UnitTest]
 public class SqlResponseTests
