+++ source/Databricks/source/SqlStatementExecution/Abstractions/ISqlStatusResponseParser.cs
@@ -12,23 +12,19 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Internal.Models;
+using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Models;
 
-namespace Energinet.DataHub.Core.Databricks.SqlStatementExecution.Internal;
+namespace Energinet.DataHub.Core.Databricks.SqlStatementExecution.Abstractions;
 
 /// <summary>
-/// Parses the response from a Databricks SQL statement execution.
+/// This interface is used to parse the response from the Databricks SQL API.
 /// </summary>
-public interface ISqlChunkDataResponseParser
+public interface ISqlStatusResponseParser
 {
     /// <summary>
-    /// Parses the chunk data response from a Databricks SQL statement execution.
-    ///
-    /// The Chunk Data can vary in response depending on the type of statement executed.
-    /// Therefore a list of column names is required to parse the response.
+    /// Parse the response from the Databricks SQL API.
     /// </summary>
     /// <param name="jsonResponse"></param>
-    /// <param name="columnNames"></param>
-    /// <returns>Returns a <see cref="TableChunk"/></returns>
-    TableChunk Parse(string jsonResponse, string[] columnNames);
+    /// <returns>Returns <see cref="SqlResponse"/></returns>
+    SqlResponse Parse(string jsonResponse);
 }
