+++ source/Databricks/source/SqlStatementExecution/Extensions/DependencyInjection/DatabricksSqlStatementExecutionExtensions.cs
@@ -14,6 +14,7 @@
 
 using System;
 using System.Net.Http.Headers;
+using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Abstractions;
 using Energinet.DataHub.Core.Databricks.SqlStatementExecution.AppSettings;
 using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Internal;
 using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Internal.Constants;
