+++ source/Databricks/source/SqlStatementExecution/Internal/SqlChunkResponseParser.cs
@@ -13,7 +13,8 @@
 // limitations under the License.
 
 using System;
-using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Internal.Models;
+using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Abstractions;
+using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Models;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
