+++ source/Databricks/source/SqlStatementExecution/Internal/SqlStatusResponseParser.cs
@@ -14,7 +14,8 @@
 
 using System;
 using System.Linq;
-using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Internal.Models;
+using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Abstractions;
+using Energinet.DataHub.Core.Databricks.SqlStatementExecution.Models;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
