+++ source/Databricks/source/SqlStatementExecution/Models/SqlResultRow.cs
@@ -12,17 +12,21 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Energinet.DataHub.Core.Databricks.SqlStatementExecution.Internal.Models;
+namespace Energinet.DataHub.Core.Databricks.SqlStatementExecution.Models;
 
 /// <summary>
-/// Representation of the state of a Databricks SQL response.
+/// This class is optimized to reduce allocations when reading data from the database.
 /// </summary>
-public enum SqlResponseState
+public record SqlResultRow
 {
-    Failed = 0,
-    Cancelled = 1,
-    Pending = 2,
-    Succeeded = 3,
-    Running = 4,
-    Closed = 5,
+    private readonly TableChunk _chunk;
+    private readonly int _index;
+
+    public SqlResultRow(TableChunk chunk, int index)
+    {
+        _chunk = chunk;
+        _index = index;
+    }
+
+    public virtual string this[string column] => _chunk[_index, column];
 }
